+++ tests/ui/apps/DatePicker/DatePicker-View.js
